Se implementó un nuevo sistema de carga utilizando el hook useLoad, mejorando la gestión de datos del usuario.
